easydeploy-server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
celery>=5.3.0
redis>=4.5.5
boto3>=1.26.0
docker>=6.1.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0 